==> domestic_hourly_data/step1_obs_hourly_preprocess.ipynb <==
#这个程序作用如下：
1.合并散文件。


import numpy as np
import os
import matplotlib.pyplot as plt
import glob
import datetime
import pandas as pd

#start='2018010100'
#end='2019010100'
#datestart=datetime.datetime.strptime(start,'%Y%m%d%H')
#dateend=datetime.datetime.strptime(end,'%Y%m%d%H')
#print(datestart.strftime('%Y%m%d%H'))
data_path='/mnt/hdd1/OBS/hourly/2018/'

sta_name=[]
lat=[]
lon=[]
ht=[]
psl=[]
tem=[]
td=[]
wdir=[]
wspd=[]
pr1=[]
pr6=[]
pr24=[]
ps=[]



mulu = glob.glob(data_path+'*')
mulu.sort()
df_all=pd.DataFrame(columns=['time', 'sta', 'ht', 'lat', 'lon', 'pr1', 'pr24', 'pr6', 'ps', 'psl', 't2m', 'td', 'wdir', 'wspd'])
#print(mulu)
df_final=pd.DataFrame(columns=['time', 'sta', 'ht', 'lat', 'lon', 'pr1', 'pr24', 'pr6', 'ps', 'psl', 't2m', 'td', 'wdir', 'wspd'])


for file in mulu:
    #print(file)
    with open(os.path.join(data_path, file)) as fdata:
        fdata.readline() #跳过第一行
        while True:
            line1=fdata.readline()
            #print(line1)
            if not line1:
                break                           
            #sta_name0,_,lat0,lon0,ht0,_,_ = [i for i in line1.split()] # 将整行数据分割处理，如果分割符是空格，括号里就不用传入参数，如果是逗号， 则传入‘，'字符。
            #由于数据有黏连的情况，只能按照字符数读取
            sta_name0=line1[0:8]
            lat0=line1[23:28]
            lon0=line1[32:38]
            ht0=line1[41:48]
            #print(sta_name0,lat0,lon0,ht0)
         
            line2=fdata.readline()
            if not line2:
                break 
            psl0,tem0,td0,wdir0,wspd0,pr10,pr60,pr240,ps0,_,_,_,_,_,_,_,_,_= [float(i) for i in line2.split()]
            sta_name.append(sta_name0.strip())  # 添加新读取的数据
            lat.append(lat0.strip())
            lon.append(lon0.strip())
            ht.append(ht0.strip())
            psl.append(psl0)
            tem.append(tem0)
            td.append(td0)
            wdir.append(wdir0)
            wspd.append(wspd0)
            pr1.append(pr10)
            pr6.append(pr60)
            pr24.append(pr240)
            ps.append(ps0)            
            pass
    sta_name = np.array(sta_name).astype('str') # 将数据从list类型转换为array类型。
    lat = np.array(lat).astype('float') 
    lon = np.array(lon).astype('float')  
    ht = np.array(ht).astype('float')  
    psl = np.array(psl).astype('float') 
    tem = np.array(tem).astype('float') 
    td = np.array(td).astype('float') 
    wdir = np.array(wdir).astype('float') 
    wspd = np.array(wspd).astype('float') 
    pr1 = np.array(pr1).astype('float') 
    pr6 = np.array(pr6).astype('float') 
    pr24 = np.array(pr24).astype('float')
    ps = np.array(ps).astype('float')
    ###把numpy转换成dataframe
    df1=pd.DataFrame({'sta':sta_name,'lat':lat,'lon':lon,'ht':ht,'psl':psl,'t2m':tem,'td':td,'wdir':wdir,'wspd':wspd,'pr1':pr1,'pr6':pr6,'pr24':pr24,'ps':ps})
    str_date=file[-26:-16]#年月日小时
    #print(str_date[-2:])

    df1.insert(0,'time', str_date)#插入时间作为标记
    df1_sta=df1['sta']#把sta这一列顺序往前调
    df1.drop('sta',axis=1,inplace=True)
    df1.insert(1,'sta', df1_sta)
    df_all=pd.concat([df_all,df1],ignore_index=True)
    del df1
    del sta_name
    del lat
    del lon
    del ht
    del psl
    del tem
    del td 
    del wdir
    del wspd
    del pr1
    del pr6
    del pr24
    del ps
    sta_name=[]
    lat=[]
    lon=[]
    ht=[]
    psl=[]
    tem=[]
    td=[]
    wdir=[]
    wspd=[]
    pr1=[]
    pr6=[]
    pr24=[]
    ps=[]


df_all.to_csv('./df_all.csv', sep=',', header=True, index=False)
