==> domestic_hourly_data/step1.5_obs_hourly_preprocess(舍弃).ipynb <==
#此程序最后被*2400的程序所取代，原因是此程序的筛选标准太严格，只能筛选出42个站点。对应的输出文件为df_final_2.csv

import numpy as np
import os
import matplotlib.pyplot as plt
import glob
import datetime
import pandas as pd

data_path='/home/liuli/GTS_GFS_MOS/domestic_hourly_data/df_all.csv'
file0= pd.read_csv(data_path,sep=',')
df0=pd.DataFrame(file0)

#print(df_all.info())
df_all=df0.drop_duplicates(subset = ['time','sta'],keep='last',inplace=False) #去掉重复的行
print(df_all)

#1.先把大df中某个站点的数据挑选出来
start='2018-01-01 00:00:00'
end_of_db = '2019-01-01 00:00:00'
date_index = pd.date_range(start, end_of_db, freq = '1h')
date_df = pd.DataFrame(date_index,columns=['time'])


#把df_all的时间转换成datatime
df_all['time']=pd.to_datetime(df_all['time'],format="%Y%m%d%H")
#df_all['time'] = df_all['time'].apply(lambda x : datetime.datetime.strptime(x,'%Y%m%d%H'))
#print(df_all['time'].dtype)

#print(type(df_all['time'].dt.hour[0]))
df_qc=df_all.drop_duplicates(subset = ['sta'],keep='last',inplace=False) #申明一个不重复站点的df
print(df_qc)

#df2=df_all[df_all['sta'].isin(['53603'])]
#print(df2)
#for i in df_qc.sta:  #站点不重复
#    #print(i)
#    #print(type(i))
#    df2=df_all[df_all['sta'].isin([i])] #先挑出这个站号的行 
#    df2=df2.sort_values(by='time')#把这个站点按照时间顺序排列    
#    df2=df2.reset_index(drop=True)
#    if len(df2)>4000:
#        print(df2)

############开始循环读取每个站点的数据，并判断此站点的有效数据量
df_final=pd.DataFrame(columns=['time', 'sta', 'ht', 'lat', 'lon', 'pr1', 'pr24', 'pr6', 'ps', 'psl', 't2m', 'td', 'wdir', 'wspd'])

for i in df_qc.sta:  #站点不重复
    
    #print(type(i))
    df2=df_all[df_all['sta'].isin([i])] #先挑出这个站号的行
    if len(df2)>=4000: ####先判断此站点的总时间长度是否达到7000（根据24*365=8760的50%有效数据）
        print(i)
        df2=df2.sort_values(by='time')#把这个站点按照时间顺序排列    
        df2=df2.reset_index(drop=True)
        #print(df2)         
        # merge日期补充完整数据
        df2 = pd.merge(date_df,df2, on=['time'], how='left')
        df2['sta'].fillna(i,inplace=True)
        #df2['sta']=df2['sta'].replace('NaN',i)
        #判断03时，处理为3小时累计降水
        for index, item in enumerate(df2['time'].values):
            #从第三行开始，把1h降水量累加成为3h降水量
            if index>=2 and df2['time'].dt.hour[index]%3==0 and df2['pr1'].values[index]<1000.0:
#            if index>=2 and df2['time'].dt.hour[index]%3==0:                
#                print(index,item,df2['sta'].values[index],df2['pr1'].values[index])
                if df2['pr1'].values[index-1]<1000.0 and df2['pr1'].values[index-2]<1000.0:            
                    df2['pr1'].values[index]=df2['pr1'].values[index]+df2['pr1'].values[index-1]+df2['pr1'].values[index-2]
                    print(index,item,df2['sta'].values[index],df2['pr1'].values[index])
                    
                else:
                    df2['pr1'].values[index]=999999.0
                    print(index,item,df2['sta'].values[index],df2['pr1'].values[index])
        df_final=pd.concat([df_final,df2],ignore_index=True)            
    #print(df2)
    #print(df2.info())
    del df2

#这种根据某一列数值，赋值另一列适合比较简单的场景：print(df_all.loc[(float(df_all.time[-2:])%3 ==0), 'pr1'] )
df_final

df_final.to_csv('./df_final_2.csv', sep=',', header=True, index=False)
