==> domestic_hourly_data/step3_retrive_target_data.ipynb <==
#1.根据statistics_valid_data.txt，针对每一个要素删选出排名前300的站点
#2.每一个要素一个文件，生成完整的3-hourly时间序列。


import numpy as np
import os
import matplotlib.pyplot as plt
import datetime
import pandas as pd
txt_path='/home/liuli/GTS_GFS_MOS/domestic_hourly_data/statistics_valid_data_0.8.txt'
file1= pd.read_csv(txt_path,sep='\t',header=None,names=['sta','para','rate'])
df1=pd.DataFrame(file1)
print(df1.info())
df1

data_path='/home/liuli/GTS_GFS_MOS/domestic_hourly_data/df_final_2400.csv'
file0= pd.read_csv(data_path,sep=',')
df0=pd.DataFrame(file0)
df0=df0.replace(999999.0, np.nan) 
df0=df0.replace(999998.0, np.nan)
df0=df0.replace(9999.0, np.nan)
df0['sta']=df0['sta'].astype('str')
#df0['sta']=df0['sta'].replace('NaN','53603')
df0['time']=pd.DatetimeIndex(df0['time'])

start='2018-01-01 00:00:00'
end_of_db = '2019-01-01 00:00:00'
date_index = pd.date_range(start, end_of_db, freq = '3h')
date_df = pd.DataFrame(date_index,columns=['time'])
df0 = pd.merge(date_df,df0, on=['time'], how='left') 


print(df0.info())
print(df0.count())
df0

#每一个要素中，按照rate排名前300来筛选站点
#附属产品：所有6个要素的站点信息，放在一个txt文档里面，为以后GFS提取数据使用
#cols=['ps','psl','t2m','td','wdir','wspd']
cols=['wspd'] #不知道为什么，只能分开输出station信息，合在一起write line会导致写的信息有缺失
print(cols)
#print(df0[df0['sta'].isin(['54578'])].dropna(subset=['lat','lon'],how='any'))
for i in cols:
    name='6-para-station-info-'+i+'.txt'
    f=open(name,'w')
    #df_final=pd.DataFrame(columns=['time','sta','lat','lon',i])
    #print(df_final)
    df2=df1[df1['para'].isin([i])] #先挑出这个要素
    #print('df2=',df2)
    df2=df2.sort_values(by='rate',ascending=False)#把这个站点按照rate顺序排列  
    df2=df2.reset_index(drop=True)
    print(df2.sta[:300])
   
    for j in df2.sta[:300]:   #只挑前300个站点
        #print(j)
        df3=df0[df0['sta'].isin([j])]
        df3=df3.reset_index(drop=True)
        #print('df3=',df3)
        df4=df3[['time','sta','lat','lon',i]]
        df5=df4.dropna(subset=['lat','lon'],how='any')
        df5=df5.reset_index(drop=True)
        #print(df5)
        #print(str(df5.loc[0,'lat']))
        line=[]
        line=str(j)+'\t'+str(df5.loc[0,'lat'])+'\t'+str(df5.loc[0,'lon'])+'\n' #站号、经纬度
        f.write(line)
        #df_final=pd.concat([df_final,df4],ignore_index=True)            
        del df3
        del df4
        del df5
        del line
    #df_final.to_csv('./df_'+i+'.csv', sep=',', header=True, index=False)    
    #del df_final
    del df2
