==> domestic_hourly_data/step4_GFS_data（被ncl程序取代）.ipynb <==
import numpy as np
import os
import matplotlib.pyplot as plt
import datetime
import pandas as pd
txt_path='/home/liuli/GTS_GFS_MOS/domestic_hourly_data/6-para-station-info.txt'
file1= pd.read_csv(txt_path,sep='\t',header=None,names=['sta','lat','lon'])
df1=pd.DataFrame(file1)
print(df1.info())
df1

import pygrib as pg
grbs = pg.open('/mnt/hdd1/GFS/2018/20180415/gfs.0p25.2018041500.f006.grib2')# 所有变量
for grb in grbs:  
	print(grb) #每一个变量的头文件
#	print(grb.keys() )  #每一个变量的keys
#	print(grb.values )  # 每一个变量的值
